All: Support float arguments with CGO_ENABLED=1 on Linux
