+++ dlfcn_nocgo_linux.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2022 The Ebitengine Authors
 
-//go:build !cgo
+//go:build !cgo || arm64 || amd64
 
 package purego
