+++ dlfcn_stubs.s
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2022 The Ebitengine Authors
 
-//go:build darwin || freebsd || (linux && !cgo)
+//go:build darwin || freebsd || (linux && (!cgo || arm64 || amd64))
 
 #include "textflag.h"
