+++ sys_arm64.s
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2022 The Ebitengine Authors
 
-//go:build darwin || freebsd || (!cgo && linux) || windows
+//go:build darwin || freebsd || linux || windows
 
 #include "textflag.h"
 #include "go_asm.h"
